backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.5.2
psycopg2-binary==2.9.9
pandas==2.2.3
watchdog==5.0.3

